backend/requirements.txt
fastapi~=0.111.0
uvicorn~=0.29.0
pydantic~=2.7.1; python_version < "3.13"
pydantic>=2.10,<3; python_version >= "3.13"
pydantic-settings~=2.2.1; python_version < "3.13"
pydantic-settings>=2.6,<3; python_version >= "3.13"
python-dotenv~=1.0.1
sqlalchemy~=2.0.29
alembic~=1.13.1
psycopg2-binary~=2.9.9
geoalchemy2~=0.15.1
celery~=5.3.6
redis~=5.0.3
lightgbm~=4.3.0
scikit-learn~=1.4.2; python_version < "3.13"
scikit-learn>=1.6,<1.8; python_version >= "3.13"
shap~=0.45.0; python_version < "3.13"
shap>=0.47,<0.49; python_version >= "3.13"
pyod~=2.0.1
statsmodels~=0.14.1
joblib~=1.3.2
pandas~=2.2.1; python_version < "3.13"
pandas>=2.2.3,<2.4; python_version >= "3.13"
numpy~=1.26.4; python_version < "3.13"
numpy>=2.1,<3; python_version >= "3.13"
geopandas~=0.14.4; python_version < "3.13"
geopandas>=1.0,<1.2; python_version >= "3.13"
shapely~=2.0.4; python_version < "3.13"
shapely>=2.0.7,<2.2; python_version >= "3.13"
networkx~=3.3
razorpay~=1.4.1
httpx~=0.27.0
requests~=2.31.0
pytest~=8.2.0
pytest-asyncio~=0.23.6
pytest-cov~=5.0.0
